+++ ModelSwapperSkins/BoneMapping/BoneInitializer.cs
@@ -74,6 +74,7 @@ namespace ModelSwapperSkins.BoneMapping
             AddCustomBoneInitializerRules(BoneInitializerRules_Parent.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_Railgunner.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_Scav.Instance);
+            AddCustomBoneInitializerRules(BoneInitializerRules_Seeker.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_Shopkeeper.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_Titan.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_Toolbot.Instance);